Updated events generator, moved BlockEventsFixture due to the circular dependency
